+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ## Features
 
+* Add techdivision/import#163
 * Add techdivision/import-cli-simple#216
 * Add techdivision/import-configuration-jms#25
 * Switch to latest techdivision/import 15.* version as dependency